Move RK tableaus to separate file
