+++ netket/experimental/dynamics/__init__.py
@@ -25,8 +25,8 @@ __all__ = [
     "RK45",
 ]
 
-from .runge_kutta import RungeKuttaIntegrator, RKIntegratorConfig
-from .runge_kutta import Euler, Heun, Midpoint, RK4, RK12, RK23, RK45
+from ._rk_solver import RungeKuttaIntegrator, RKIntegratorConfig
+from ._rk_solver import Euler, Heun, Midpoint, RK4, RK12, RK23, RK45
 
 from netket.utils import _hide_submodules
